Updated example with new property

+++ pypsq_examples.py
@@ -9,16 +9,17 @@ datapass = "qqqqqQ!1"
 psql = pypsql.PyPSQL(dbname=database, dbuser=datauser, dbpass=datapass)
 print(psql.dsn_string())
 psql.connect()
-a = psql.query(query="SELECT * FROM vacancies LIMIT 1;")
-for r in a.fetchall():
-    print(r[0])
-psql.create_table(table_name="IDs", table_schema="ID NUMERIC")
-psql.query(query="INSERT INTO IDs SELECT vac_id FROM vacancies;")
-a = psql.query(query="SELECT count(*) FROM IDs;")
-print(str(a.fetchone()[0]))
-psql.truncate(table_name="IDs")
-psql.rename_table(old_table_name="IDs", new_table_name="IDs_new")
-a = psql.check_table_exist(table_name="IDs")
-print(str(a))
-psql.drop_table(table_name="IDs_new")
-del psql
+if psql.connected:
+    a = psql.query(query="SELECT * FROM vacancies LIMIT 1;")
+    for r in a.fetchall():
+        print(r[0])
+    psql.create_table(table_name="IDs", table_schema="ID NUMERIC")
+    psql.query(query="INSERT INTO IDs SELECT vac_id FROM vacancies;")
+    a = psql.query(query="SELECT count(*) FROM IDs;")
+    print(str(a.fetchone()[0]))
+    psql.truncate(table_name="IDs")
+    psql.rename_table(old_table_name="IDs", new_table_name="IDs_new")
+    a = psql.check_table_exist(table_name="IDs")
+    print(str(a))
+    psql.drop_table(table_name="IDs_new")
+    del psql
